- Nest component under parent- Fixup some styling based on mocks
